+++ src/main/java/com/aetherteam/enhanced_extinguishing/data/generators/tags/ExtinguishingBlockTagData.java
@@ -5,8 +5,8 @@ import com.aetherteam.enhanced_extinguishing.block.ExtinguishingBlocks;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.data.PackOutput;
 import net.minecraft.tags.BlockTags;
-import net.minecraftforge.common.data.BlockTagsProvider;
-import net.minecraftforge.common.data.ExistingFileHelper;
+import net.neoforged.neoforge.common.data.BlockTagsProvider;
+import net.neoforged.neoforge.common.data.ExistingFileHelper;
 
 import javax.annotation.Nullable;
 import java.util.concurrent.CompletableFuture;
@@ -16,7 +16,6 @@ public class ExtinguishingBlockTagData  extends BlockTagsProvider {
         super(output, registries, EnhancedExtinguishing.MODID, helper);
     }
 
-    @SuppressWarnings("unchecked")
     @Override
     public void addTags(HolderLookup.Provider provider) {
         this.tag(BlockTags.WALL_POST_OVERRIDE).add(ExtinguishingBlocks.EXTINGUISHED_TORCH.get());